fix error on evaluation of top level node
Before, if the cursor was on an empty line or a semicolon and thus on
the julia top level node "source_file" while evaluating an error would
be thrown. As far as I can tell it's not trivial to make the cursor go
to the next valid node, as the top level node is not a normal sibling
node. Best I could do was to print an error to let the user know
evaluation did not work because of the top level node and if the cursor
was on a semicolon, the cursor now jumps to the first column so that a
valid evaluation can follow. Neither evaluation on empty lines nor on
semicolons now throws an error.